Moved DSRBalanceCounter to own component
