+++ src/hooks/useWalletBalances.js
@@ -1,4 +1,4 @@
-import { useMemo, useState } from 'react';
+import { useMemo } from 'react';
 import useStore from 'hooks/useStore';
 import useMaker from 'hooks/useMaker';
 import { getSavingsBalance } from 'reducers/accounts';
